backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
google-genai
google-generativeai
pydantic
python-multipart
sqlalchemy